+++ src/main/kotlin/kotools/types/number/NonZeroInt.kt
@@ -1,5 +1,12 @@
 package kotools.types.number
 
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.descriptors.PrimitiveKind
+import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
+import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
 import kotools.types.annotations.SinceKotoolsTypes
 import kotools.types.string.NotBlankString
 
@@ -78,6 +85,7 @@ public fun String.toNonZeroIntOrNull(): NonZeroInt? =
  * [IllegalArgumentException] if the [value] equals `0`.
  */
 @JvmInline
+@Serializable(NonZeroIntSerializer::class)
 @SinceKotoolsTypes("1.1")
 public value class NonZeroInt
 @Throws(IllegalArgumentException::class)
@@ -372,4 +380,18 @@ public constructor(public val value: Int) : Comparable<Int> {
             null
         }
     }
+}
+
+@SinceKotoolsTypes("2.1")
+internal object NonZeroIntSerializer : KSerializer<NonZeroInt> {
+    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor(
+        NonZeroInt::class.simpleName ?: "AnonymousNonZeroInt",
+        PrimitiveKind.INT
+    )
+
+    override fun serialize(encoder: Encoder, value: NonZeroInt): Unit =
+        encoder.encodeInt(value.value)
+
+    override fun deserialize(decoder: Decoder): NonZeroInt =
+        decoder.decodeInt().toNonZeroInt()
 }