+++ changelog.md
@@ -22,6 +22,8 @@ All notable changes to this project will be documented in this file.
 - Conversions from `String` and `NotBlankString` types to `NonZeroInt`,
   `PositiveInt`, `StrictlyPositiveInt`, `NegativeInt` and `StrictlyNegativeInt`
   ([#48](https://github.com/kotools/types/issues/48)).
+- Serialization and deserialization of `NonZeroInt`
+  ([#65](https://github.com/kotools/types/issues/65)).
 
 ## Version 2.0.0 - 2022/08/01